Extract FieldGroup component
